requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
asyncpg==0.29.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.0
fastapi-cli==0.0.5
fastjsonschema==2.20.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-bcrypt==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.11
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.6
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
yarg==0.1.9
yarl==1.13.1
starlette-admin==0.14.1
sqlmodel==0.0.22
sqlalchemy==2.0.36
psycopg2==2.9.10
sqlalchemy_file==0.6.0
pillow==11.0.0
fasteners==0.19
apache-libcloud==3.8.0